removed circular imports for python 2
