+++ mlflow/spark.py
@@ -345,7 +345,7 @@ def load_model(path, run_id=None, dfs_tmpdir=None):
     >>> prediction = model.transform(test)
     """
     if run_id is not None:
-        path = mlflow.tracking.utils._get_model_log_dir(model_name=path, run_id=run_id)
+        path = mlflow.tracking.artifact_utils._get_model_log_dir(model_name=path, run_id=run_id)
     path = os.path.abspath(path)
     flavor_conf = _get_flavor_configuration(model_path=path, flavor_name=FLAVOR_NAME)
     spark_model_artifacts_path = os.path.join(path, flavor_conf['model_data'])
