+++ tests/pyfunc/test_model_export_with_class_and_artifacts.py
@@ -22,8 +22,8 @@ import mlflow.pyfunc.scoring_server as pyfunc_scoring_server
 import mlflow.sklearn
 from mlflow.exceptions import MlflowException
 from mlflow.models import Model
-from mlflow.tracking.utils import get_artifact_uri as utils_get_artifact_uri
-from mlflow.tracking.utils import _get_model_log_dir
+from mlflow.tracking.artifact_utils import get_artifact_uri as utils_get_artifact_uri
+from mlflow.tracking.artifact_utils import _get_model_log_dir
 from mlflow.utils.environment import _mlflow_conda_env
 from mlflow.utils.model_utils import _get_flavor_configuration
